It continues!
Added display of albums. For now, the AJAX loads the page with full <img> tags. Thereafter, it will fetch a list of images that will be loaded one by one.

Small revisions in the design. For now, everything is basic, we will make it all nice at the end.

Changing the cron job: The proportions of the image are now preserved when creating optimized images. Similar work would be done to the thumbnails. But I'd like to do more than that a crop of 100x100, but after resizing. To do that too!

Anyway, the library Imagine really works! But the documentation is rather light and requires to watch in classes in the Github repository to know how objects work. A good PHP Doc would not hurt!