Added : photo.php to display one photo (with his photo.tpl).
Updated : style.css to add new div (photo)
Updated : album.php (and album.tpl) : the directory of each photo was added in the view (to request displaying of one photo via AJAX)

In a next commit : Modify style when a photo is displayed. Add more informations (exif info, download button, etc…).
And… review all style of this application. It's time to do a clean and modern style :)