Updated : album.php and functions.js.
In album, we're getting all photos in directories. We check if the photo exist in a compressed version too. The system try to load the compressed photo instead full resolution photo to reduce brand with consumption.
But, if the compressed photo isn't available (where is the cron ? :-( ), we take the full resolution (it can be very very long…)

In functions.js, adding support for changing title of the browser window. It will be used when you're in an album or in a photo.

That's all for tonight, folks !